Non-functional improvements to Kruskal
